+++ RevitAddin.Icon.Example/Revit/App.cs
@@ -2,6 +2,7 @@ using Autodesk.Revit.DB;
 using Autodesk.Revit.UI;
 using ricaun.Revit.UI;
 using System.IO;
+using System.Linq;
 
 namespace RevitAddin.Icon.Example.Revit
 {
@@ -14,10 +15,37 @@ namespace RevitAddin.Icon.Example.Revit
             var icons = new[] { "Grey", "Red", "Orange", "Yellow", "Green", "Cyan", "Blue", "Purple", "Pink", "Brown" };
 
             ribbonPanel = application.CreatePanel("RevitAddin.Icon.Example");
+
             foreach (var icon in icons)
             {
                 ribbonPanel.CreatePushButton<Commands.Command>(icon)
-                    .SetLargeImage($"Resources/Images/Cube-{icon}-Light.ico");
+                    .SetLargeImage($"Resources/Images/Cube-{icon}-Light.tiff");
+            }
+
+            ribbonPanel.AddSlideOut();
+
+            foreach (var icon in icons)
+            {
+                ribbonPanel.RowLargeStackedItems(
+                    ribbonPanel.CreatePushButton<Commands.Command>(icon)
+                        .SetLargeImage($"Resources/Images/Cube-{icon}-Light.tiff")
+                        .SetShowText(),
+                    ribbonPanel.CreatePushButton<Commands.Command>(icon)
+                        .SetLargeImage($"Resources/Images/Box-{icon}-Light.tiff")
+                        .SetShowText()
+                );
+            }
+
+            foreach (var icon in icons)
+            {
+                ribbonPanel.RowStackedItems(
+                    ribbonPanel.CreatePushButton<Commands.Command>(icon)
+                        .SetLargeImage($"Resources/Images/Cube-{icon}-Light.tiff")
+                        .SetShowText(),
+                    ribbonPanel.CreatePushButton<Commands.Command>(icon)
+                        .SetLargeImage($"Resources/Images/Box-{icon}-Light.tiff")
+                        .SetShowText()
+                );
             }
 
             return Result.Succeeded;